// src/lib/interfaces/admin-interface/userRights.ts
export interface UserRightsInterface {
  id?: number;
  group_code?: string | string[] | undefined;
  pos_allow_manual_memo_disc?: boolean;
  pos_show_total_sales_amount?: boolean;
  pos_allow_hold_memo?: boolean;
  pos_allow_item_disc?: boolean;
  pos_allow_promo_item_return?: boolean;
  pos_allow_drawer_view?: boolean;
  pos_allow_lookup?: boolean;
  pos_allow_customer_ledger?: boolean;
  pos_allow_tender_amount?: boolean;
  pos_allow_order_tender_amount?: boolean;
  pos_allow_negative_memo?: boolean;
  pos_allow_non_barcoded_price_changes?: boolean;
  pos_generate_serial_no_userwise?: boolean;
  pos_allow_authorization?: boolean;
  pos_allow_rate_change?: boolean;
  pos_allow_service_memo_rate_change?: boolean;
  pos_allow_advance_print_receipt?: boolean;
  pos_allow_stockpoint_changes?: boolean;
  pos_allow_multi_salesman_tagging?: boolean;
  pos_allow_salesman_tagging?: boolean;
  pos_round_off?: number | string;
  pos_cash_memo_copies?: number | null;
  pos_max_disc_percent?: number | null;
  pos_max_disc_amount?: number | null;
  pos_cm_back_date_days?: number | null;
  pos_price_override?: boolean;
  pos_return_memo_selection?: number | string;
  pos_allow_credit_advance_system?: boolean;
  pos_allow_promoview?: boolean;
  pos_return_memo_mode?: number | string;
  pos_return_memo_copies?: number | null;
  pos_rm_back_date_days?: number | null;
  pos_return_memo_for_cm_allowed_days?: number | null;
  pos_return_adj_allowed_days?: number;
  pos_non_sale_item?: boolean;
  stockpoint_id?: number | string;
  time_stamp?: string;
  pos_teller_summary?: boolean;
  pos_cancel_ds?: boolean;
  pos_cancel_cash_memo?: boolean;
  pos_cancel_return_memo?: boolean;
  pos_modify_ds?: boolean;
  pos_modify_cash_memo?: boolean;
  pos_modify_return_memo?: boolean;
  pos_sales_margin_percent?: number;
  pos_exclusive_of_tax?: boolean;
  pos_allow_multi_price?: boolean;
  pos_allow_multi_currency?: boolean;
  pos_reprint_days?: number;
  pos_color1?: string;
  pos_color2?: string;
  pos_color3?: string;
  pos_color_apply?: boolean;
  pos_ds_color1?: string;
  pos_ds_color2?: string;
  pos_ds_color3?: string;
  pos_ds_color_apply?: boolean;
  pos_against_cash_color1?: string;
  pos_against_cash_color2?: string;
  pos_against_cash_color3?: string;
  pos_against_cash_color_apply?: boolean;
  pos_posr_color1?: string;
  pos_posr_color2?: string;
  pos_posr_color3?: string;
  pos_posr_color_apply?: boolean;
  pos_member_no_edit?: boolean;
  pos_member_tagging?: boolean;
  pos_ret_ref_mandatory?: boolean;
  pos_return_both_way?: boolean;
  inv_category1?: boolean;
  inv_category2?: boolean;
  inv_category3?: boolean;
  inv_category4?: boolean;
  inv_category5?: boolean;
  inv_category6?: boolean;
  fin_post?: boolean;
  inv_category7?: boolean;
  inv_category8?: boolean;
  inv_category9?: boolean;
  inv_category10?: boolean;
  inv_category11?: boolean;
  inv_category12?: boolean;
  inv_category13?: boolean;
  inv_category14?: boolean;
  inv_category15?: boolean;
  inv_category16?: boolean;
  inv_dept_code?: boolean;
  inv_set_code?: boolean;
  inv_item_code?: boolean;
  inv_customize?: boolean;
  ws_dc_price_edit?: boolean;
  rep_period_restriction?: boolean;
  rep_period_days?: number;
  pos_points_voucher?: boolean;
  inv_allow_margin_rule_edit?: boolean;
  inv_allow_eway_mandatory?: boolean;
  pos_print_based_on?: number;
  proc_restrict_grn?: boolean;
  proc_restrict_dc?: boolean;
  proc_restrict_pi?: boolean;
  proc_restrict_si?: boolean;
  proc_restrict_dn?: boolean;
  proc_restrict_cn?: boolean;
  pos_block_member_create?: string | number;
  pos_max_less_qty?: number;
  pos_allow_wsp?: boolean;
  pos_wsp_validation_mandatory?: boolean;
  pos_reprint_count?: number;
  ws_salesman_mandatory?: boolean;
  rep_allow_mw_dashboard?: boolean;
  pos_max_amount?: number;
  inv_out_stkpt_list?: string | number;
  inv_allow_receive_all?: boolean;
  msd_allow_receive_all?: boolean;
  inv_show_inventory_cat?: boolean;
  company_id?: number;
  branch_id?: number;
  fin_back_entry_days?: number;
  pos_allow_gift_voucher_manualentry?: boolean;
  pos_max_refund_amount?: number;
  [key: string]: any;
  allow_catalogue?: boolean;
}